+++ sentry-sidekiq/spec/sentry/sidekiq/sentry_context_middleware_spec.rb
@@ -3,82 +3,52 @@ require "spec_helper"
 RSpec.shared_context "sidekiq", shared_context: :metadata do
   let(:user) { { "id" => rand(10_000) } }
 
-  let(:client) do
-    Sidekiq::Client.new.tap do |client|
-      client.middleware do |chain|
-        chain.add Sentry::Sidekiq::SentryContextClientMiddleware
-      end
-    end
+  let(:processor) do
+    options = { queues: ['default'] }
+    Sidekiq::Manager.new(options).workers.first
   end
 
-  let(:random_empty_queue) do
-    Sidekiq::Queue.new(rand(10_000)).tap do |queue|
-      queue.clear
-    end
+  let(:transport) do
+    Sentry.get_current_client.transport
   end
 end
 
 RSpec.describe Sentry::Sidekiq::SentryContextServerMiddleware do
   include_context "sidekiq"
 
-  after { Sidekiq::RetrySet.new.clear }
-
-  it "sets user from the job to events" do
-    perform_basic_setup { |config| config.traces_sample_rate = 1.0 }
-    Sentry.set_user(user)
-
-    queue = random_empty_queue
-    options = { queues: [queue.name] }
-    processor = Sidekiq::Manager.new(options).workers.first
-
-    client.push('queue' => queue.name, 'class' => HappyWorker, 'args' => [])
-
-    expect { processor.send(:process_one) }.
-      to change { Sentry.get_current_client.transport.events.size }.by(1)
-
-    event =  Sentry.get_current_client.transport.events.first
-    expect(event).not_to be_nil
-    expect(event.user).to eq(user)
-  end
-
-  it "sets user from the job to sidekiq event if worker raises an exception" do
+  it "sets user to the event" do
     perform_basic_setup { |config| config.traces_sample_rate = 0 }
     Sentry.set_user(user)
 
-    queue = random_empty_queue
-    options = { queues: [queue.name] }
-    processor = Sidekiq::Manager.new(options).workers.first
+    process_job(processor, "SadWorker")
 
-    client.push('queue' => queue.name, 'class' => SadWorker, 'args' => [])
-
-    expect do
-      begin; processor.send(:process_one); rescue RuntimeError; end
-    end.
-      to change { Sentry.get_current_client.transport.events.size }.by(1)
-
-    Sentry.get_current_client.transport.events.each do |event|
-      expect(event.user).to eq(user)
-    end
+    expect(transport.events.count).to eq(1)
+    event = transport.events.first
+    expect(event.user).to eq(user)
   end
 
-  it "sets user from the job to sidekiq and error events if worker raises an exception when trace enabled" do
-    perform_basic_setup { |config| config.traces_sample_rate = 1.0 }
-    Sentry.set_user(user)
+  context "with tracing enabled" do
+    before do
+      perform_basic_setup { |config| config.traces_sample_rate = 1.0 }
+      Sentry.set_user(user)
+    end
 
-    queue = random_empty_queue
-    options = { queues: [queue.name] }
-    processor = Sidekiq::Manager.new(options).workers.first
+    it "sets user to the transaction" do
+      process_job(processor, "HappyWorker")
 
-    client.push('queue' => queue.name, 'class' => SadWorker, 'args' => [])
+      expect(transport.events.count).to eq(1)
+      transaction = transport.events.first
+      expect(transaction).not_to be_nil
+      expect(transaction.user).to eq(user)
+    end
 
-    # XXX: In ruby 2.4, two events are pushed. In other versions, only one
-    # event is pushed. Use by_at_least.
-    expect do
-      begin; processor.send(:process_one); rescue RuntimeError; end
-    end.
-      to change { Sentry.get_current_client.transport.events.size }.by_at_least(1)
+    it "sets user to both the event and transaction" do
+      process_job(processor, "SadWorker")
 
-    Sentry.get_current_client.transport.events.each do |event|
+      expect(transport.events.count).to eq(2)
+      transaction = transport.events.first
+      expect(transaction.user).to eq(user)
+      event = transport.events.last
       expect(event.user).to eq(user)
     end
   end
@@ -87,21 +57,33 @@ end
 RSpec.describe Sentry::Sidekiq::SentryContextClientMiddleware do
   include_context "sidekiq"
 
-  before { perform_basic_setup }
+  let(:client) do
+    Sidekiq::Client.new.tap do |client|
+      client.middleware do |chain|
+        chain.add described_class
+      end
+    end
+  end
 
-  it "does not user to the job if user is absence in the current scope" do
-    queue = random_empty_queue
-    client.push('queue' => queue.name, 'class' => HappyWorker, 'args' => [])
+  # the default queue
+  let!(:queue) { Sidekiq::Queue.new("default") }
+
+  before do
+    queue.clear
+    perform_basic_setup
+  end
+
+  it "does not add user to the job if user is absence in the current scope" do
+    client.push('queue' => 'default', 'class' => HappyWorker, 'args' => [])
 
     expect(queue.size).to be(1)
     expect(queue.first["sentry_user"]).to be_nil
   end
 
   it "sets user of the current scope to the job if present" do
-    queue = random_empty_queue
     Sentry.set_user(user)
 
-    client.push('queue' => queue.name, 'class' => HappyWorker, 'args' => [])
+    client.push('queue' => 'default', 'class' => HappyWorker, 'args' => [])
 
     expect(queue.size).to be(1)
     expect(queue.first["sentry_user"]).to eq(user)